+++ packages/types/src/PagedResults.ts
@@ -1,7 +1,9 @@
+import type { Metadata } from "./Metadata.js"
+
 export interface PagedResults<T> {
   items: T[]
   /// Set this to undefined to tell the app that there are no more items
-  metadata?: unknown
+  metadata?: Metadata
 }
 
 export const EndOfPageResults: PagedResults<never> = Object.freeze({