+++ packages/types/src/impl/index.ts
@@ -9,4 +9,4 @@ export * from './CloudflareError.js'
 export * from './CookieStorageInterceptor.js'
 export * from './URL.js'
 export * from './SourceInfo.js'
-// export * from './TestDefinition.js'
+