+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @paperback/types
 
+## 1.0.0-alpha.75
+
+### Patch Changes
+
+- search filter overhaul
+
 ## 1.0.0-alpha.74
 
 ### Patch Changes